requirements.txt
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
idna==3.8
importlib_resources==6.4.4
Jinja2==3.1.4
kaggle==1.6.17
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opendatasets==0.1.22
packaging==24.1
pillow==10.4.0
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
sympy==1.13.2
text-unidecode==1.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
zipp==3.20.1
